31st commit Changes: 1) Fixed the account deletion process. Now all transactions linked to the deleted account become frozen. The account number is changed to zero, to avoid confusions with new account having same number. 2) Add JavaScript to hide buttons update and delete, and to change background color for frozen transactions. 3) Minor changes

+++ account-service/src/main/java/com/bank/accountservice/service/AccountServiceImpl.java
@@ -45,8 +45,15 @@ public class AccountServiceImpl implements AccountService {
     @Override
     @Transactional
     public void deleteAccount(Long accountId) {
+        Account account = repository.findById(accountId).orElse(null);
+        if (account == null) {
+            // TODO return message to the web-service
+            log.error("Account with id: {} not found", accountId);
+            throw new EntityNotFoundException("Account with id " + accountId + " not found");
+        }
+        Long accountNumber = account.getAccountNumber();
         repository.deleteById(accountId);
-        publisher.publishAccountDeletedEvent(accountId);
+        publisher.publishAccountDeletedEvent(accountId, accountNumber);
         log.info("Account with id: {} deleted", accountId);
     }
